RASS-1734-success message added for hearing deletion (#1248)
* RASS-1734-success message added for hearing deletion

* RASS-1734-page name renamed

* RASS-1734-page name renamed

* RASS-1734-fixing tests

* RASS-1734-small refactoring

+++ server/services/remandAndSentencingService.ts
@@ -1,5 +1,7 @@
 import type { CourtAppearance, CourtCase, Offence, UploadedDocument } from 'models'
 import { Dayjs } from 'dayjs'
+import type { DeleteHearingForm } from 'forms'
+import validate from '../validation/validation'
 import {
   ConsecutiveChainValidationRequest,
   CourtCaseCountNumbers,
@@ -99,8 +101,30 @@ export default class RemandAndSentencingService {
     return this.remandAndSentencingApiClient.getCourtAppearanceByAppearanceUuid(appearanceUuid, username)
   }
 
-  async deleteCourtAppearance(appearanceUuid: string, username: string): Promise<void> {
-    return this.remandAndSentencingApiClient.deleteCourtAppearance(appearanceUuid, username)
+  async deleteCourtAppearance(
+    appearanceUuid: string,
+    username: string,
+    deleteHearingForm: DeleteHearingForm,
+  ): Promise<
+    {
+      text?: string
+      html?: string
+      href: string
+    }[]
+  > {
+    const errors = validate(
+      deleteHearingForm,
+      {
+        deleteHearing: 'required',
+      },
+      {
+        'required.deleteHearing': 'You must select an option.',
+      },
+    )
+    if (deleteHearingForm.deleteHearing === 'true') {
+      await this.remandAndSentencingApiClient.deleteCourtAppearance(appearanceUuid, username)
+    }
+    return errors
   }
 
   async getCourtCaseDetails(courtCaseUuid: string, username: string): Promise<PageCourtCaseContent> {
